+++ chessboard/tests/test_chessboard.py
@@ -28,11 +28,10 @@ from chessboard import Chessboard, Board
 class TestChessboard(unittest.TestCase):
 
     def test_instanciation(self):
-        board = Chessboard(3, 3, king=2)
+        board = Chessboard(3, 3, king=2, queen=7)
         self.assertEquals(board.length, 3)
         self.assertEquals(board.height, 3)
-        self.assertDictContainsSubset({'king': 2}, board.pieces)
-
+        self.assertDictContainsSubset({'king': 2, 'queen': 7}, board.pieces)
 
 
 class TestBoard(unittest.TestCase):